doc: Rework overriding docs
Having overriding be a flake template was always a bit weird.
Move all overriding examples to `doc/`.

Add the internal testing build-system overrides as an example.
This overriding pattern is actually quite good and deserves to be
exposed as a way to deal with the issue of the lock file missing build systems.